+++ threedi_models_and_simulations/api_calls/threedi_calls.py
@@ -54,6 +54,7 @@ from threedi_api_client.openapi import (
     SchematisationRevision,
     Simulation,
     SimulationSettingsOverview,
+    SimulationStatus,
     SqliteFileUpload,
     TableStructureControl,
     Template,
@@ -125,6 +126,11 @@ class ThreediCalls:
     def __init__(self, threedi_api: ThreediApi) -> None:
         self.threedi_api = threedi_api
 
+    @property
+    def expiration_date(self):
+        created__date__gt = self.EXPIRATION_TIME.strftime("%Y-%m-%d")
+        return created__date__gt
+
     def paginated_fetch(self, api_method: Callable, *args, **kwargs) -> List[Any]:
         """Method for iterative fetching of the data via given API endpoint."""
         limit = self.FETCH_LIMIT
@@ -151,8 +157,9 @@ class ThreediCalls:
 
     def fetch_simulations(self) -> List[Simulation]:
         """Fetch all simulations available for current user."""
-        created__date__gt = self.EXPIRATION_TIME.strftime("%Y-%m-%d")
-        simulations_list = self.paginated_fetch(self.threedi_api.simulations_list, created__date__gt=created__date__gt)
+        simulations_list = self.paginated_fetch(
+            self.threedi_api.simulations_list, created__date__gt=self.expiration_date
+        )
         return simulations_list
 
     def fetch_simulation(self, simulation_pk: int) -> Simulation:
@@ -208,6 +215,12 @@ class ThreediCalls:
         current_status = self.threedi_api.simulations_status_list(str(simulation_pk), limit=self.FETCH_LIMIT)
         return current_status
 
+    def fetch_simulation_statuses(self, **params) -> List[SimulationStatus]:
+        """Fetch simulations statuses."""
+        params["created__date__gt"] = self.expiration_date
+        statuses = self.paginated_fetch(self.threedi_api.statuses_list, **params)
+        return statuses
+
     def fetch_simulation_progress(self, simulation_pk: int) -> Progress:
         """Get a given simulation progress. Available only if simulation was already started."""
         logger.debug("Fetching simulation progress for sim id %s...", str(simulation_pk))
@@ -232,7 +245,7 @@ class ThreediCalls:
             status_time = current_status.time
             if status_time is None:
                 status_time = 0
-            if status_name == "initialized":
+            if status_name == "initialized" and status_time:
                 sim_progress = self.threedi_api.simulations_progress_list(spk_str, limit=self.FETCH_LIMIT)
             elif status_name == "postprocessing" or status_name == "finished":
                 sim_progress = Progress(percentage=100, time=status_time)